first pass at massive file restructuring
