+++ app/common/slack-api/profiles.ts
@@ -1,6 +1,6 @@
 import { WebClient, WebAPICallResult } from '@slack/client';
 import { SSM } from 'aws-sdk';
-import { getAccessTokenParameterPath } from '../auth-redirect';
+import { getAccessTokenParameterPath } from '../../app-authorization/auth-redirect';
 
 const slack = new WebClient();
 const ssm = new SSM({ apiVersion: '2014-11-06' });
@@ -21,7 +21,7 @@ export async function getUserProfile(user: string, teamId: string): Promise<Prof
   };
 }
 
-export async function getToken(teamId: string): Promise<string> {
+async function getToken(teamId: string): Promise<string> {
   const params = {
     Name: getAccessTokenParameterPath(teamId),
     WithDecryption: true
