+++ app/common/slack-messages/initiative-details.ts
