+++ app/slash-commands/add-initiative.ts
@@ -1,9 +1,9 @@
 import { DynamoDB } from 'aws-sdk';
 import { apiWrapper, ApiSignature } from '@manwaring/lambda-wrapper';
-import { CreateInitiativeRequest, InitiativeResponse, InitiativeRecord, INITIATIVE_TYPE } from './initiative';
-import { getUserProfile } from './slack/profile';
-import { DetailResponse } from './slack-responses/initiative-details';
-import { MemberResponse, MEMBER_TYPE, getTeamIdentifier } from './member';
+import { CreateInitiativeRequest, InitiativeResponse, InitiativeRecord, INITIATIVE_TYPE } from '../initiative';
+import { getUserProfile } from '../common/slack-api';
+import { DetailResponse } from '../common/slack-messages';
+import { MemberResponse, MEMBER_TYPE, getTeamIdentifier } from '../member';
 import { SlashCommandBody } from 'slack';
 
 const initiatives = new DynamoDB.DocumentClient({ region: process.env.REGION });
