+++ serverless.yml
@@ -93,13 +93,13 @@ custom:
 
 functions:
   initiate-update-request:
-    handler: app/initiate-update-request.handler
+    handler: app/status/updates/initiate-update-request.handler
     events:
       - schedule: cron(0 15 ? * MON *) # 3PM UTC (10 AM EST) every Monday
       # - schedule: cron(0/5 * * * ? *) # every 5 minutes for testing
 
   request-update:
-    handler: app/request-update.handler
+    handler: app/status-updates/request-update.handler
     events:
       - sns: ${self:custom.requestUpdateTopic}
 
@@ -118,7 +118,7 @@ functions:
           path: list
 
   action-endpoint:
-    handler: app/action-endpoint.handler
+    handler: app/interactivity/action-endpoint.handler
     events:
       - http:
           method: post
@@ -132,7 +132,7 @@ functions:
           arn: { Fn::GetAtt: [InitiativesTable, StreamArn] }
 
   notify-when-join:
-    handler: app/notify-when-join.handler
+    handler: app/notifications/notify-when-join.handler
     events:
       - sns: ${self:custom.notifyWhenJoinTopic}
